split up table module
